LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/test_readwrite.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>window.performance is read/write</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
	<link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute"/>
	<meta name="assert" content="The window.performance attribute provides a hosting area for performance related attributes. "/>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script id="metadata_cache">/*
{
  "window.performance is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  },
  "window.performance is read/write": {},
  "var performance is read/write": {}
}
*/</script>
        <link rel="stylesheet" href="/w3c/resources/testharness.css" />
    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates that the window.performance object is read/write.</p>
        <div id="log"></div>
        <script>
            test_namespace();

            window.performance = 'foo';
            test_equals(window.performance, 'foo', 'window.performance is read/write');

            var performance = 'bar';
            test_equals(performance, 'bar', 'var performance is read/write');
        </script>
    </body>
</html>
