third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/w3c/webperf/approved/navigation-timing/html/test_readwrite-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance is read/write 
PASS var performance is read/write 
Harness: the test ran to completion.
